2018/tutorial-3/src/main/java/id/ac/ui/cs/advprog/tutorial3/composite/Company.java
package id.ac.ui.cs.advprog.tutorial3.composite;

import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

public class Company {
    protected List<Employees> employeesList;

    public Company() {
        employeesList = new ArrayList<Employees>();
    }

    public Company(List<Employees> employeesList) {
        Collections.copy(this.employeesList, employeesList);
    }

    public void addEmployee(Employees employees) {
        //TODO Implement
    }

    public double getNetSalaries() {
        //TODO Implement
    }

    public List<Employees> getAllEmployees() {
        //TODO Implement
    }
}
